refactor(ui): split apart app branch UI & add stories
