+++ services/dashboard-ui/client/components/branches/WorkflowStepDetail/steps/BuildStep/index.ts
@@ -1,0 +1,1 @@
+export { BuildStep } from './BuildStep'
