+++ services/dashboard-ui/client/components/branches/WorkflowStepDetail/steps/CommitStep/index.ts
@@ -1,0 +1,1 @@
+export { CommitStep } from './CommitStep'
