+++ services/dashboard-ui/client/components/branches/WorkflowStepDetail/steps/ConfigStep/index.ts
@@ -1,0 +1,2 @@
+export { ConfigStepContainer as ConfigStep } from './ConfigStepContainer'
+export { ConfigStep as ConfigStepComponent } from './ConfigStep'
