+++ services/dashboard-ui/client/components/branches/WorkflowStepDetail/steps/DeployGroupStep/index.ts
@@ -1,0 +1,1 @@
+export { DeployGroupStep } from './DeployGroupStep'
