+++ services/dashboard-ui/client/components/branches/WorkflowStepDetail/steps/PlanGroupStep/index.ts
@@ -1,0 +1,2 @@
+export { PlanGroupStepContainer as PlanGroupStep } from './PlanGroupStepContainer'
+export { PlanGroupStep as PlanGroupStepComponent } from './PlanGroupStep'
